refactor: extract the source study citation into its own component (#3218)
Filenamecell guarded the citation on the split result rather than on
publicationstring, so the guard read indirectly, typescript stopped narrowing
the string the no-doi branch renders, and the split ran on rows that never used
it. A citation component taking a required publicationstring puts all three
right: the caller owns whether there is a citation, the string arrives narrowed,
and the split sits in the doi branch that consumes it. The citation-only pieces
move with it, so filenamecell is now a stack and two lines.
