+++ components/HCABioNetworks/Network/Atlas/components/SourceDatasets/components/MainColumn/components/table/components/FileNameCell/components/Citation/types.ts
@@ -1,0 +1,4 @@
+export interface Props {
+  doi: string | null;
+  publicationString: string;
+}
